refactor(terminal): split inputEditorExtensions into per-chip modules
- Replace 1,299-line monolith with 8 focused modules in inputEditorExtensions/ directory
- Create base.ts with shared theme, voice fields, auto-size, keymaps, and utilities
- Create per-chip modules: slashChip, fileChip, imageChip, fileDropChip, diffChip, terminalChip, selectionChip
- Add barrel index.ts re-exporting all public symbols for backwards-compatible imports
- Share TERMINAL_ICON_SVG via base.ts to avoid cross-chip imports
